+++ backend/data/models.sql
@@ -1,13 +1,28 @@
--- name: ControllableUnitLookup :many
+-- no function as AP is public information
+-- name: GetAccountingPointIDFromBusinessID :one
+SELECT ap.id
+FROM accounting_point AS ap
+WHERE ap.business_id = @accounting_point_business_id;
+
+-- name: GetCurrentControllableUnitAccountingPoint :one
 SELECT
-    id::bigint,
-    business_id::text,
-    name::text,
     accounting_point_id::bigint,
-    end_user_id::bigint,
-    technical_resources::jsonb
+    accounting_point_business_id::text
+FROM current_controllable_unit_accounting_point(
+    @controllable_unit_business_id::text
+);
+
+-- name: ControllableUnitLookupCheckEndUserMatchesAccountingPoint :one
+SELECT end_user_id::bigint
+FROM controllable_unit_lookup_check_end_user_matches_accounting_point(
+    @end_user_business_id::text,
+    @accounting_point_business_id::text
+);
+
+-- name: ControllableUnitLookup :one
+SELECT
+    controllable_units::jsonb
 FROM controllable_unit_lookup(
-  @end_user_business_id,
   -- empty strings considered as missing values
   nullif(@controllable_unit_business_id::text, ''),
   nullif(@accounting_point_id::text, '')