+++ docs/changelog.md
@@ -4,6 +4,14 @@ This changelog tracks all changes that are visible in our API.
 
 <!-- markdownlint-disable MD013 -->
 
+## 01.07.2025
+
+* **Changed the body format of the response for the _Controllable Unit Lookup_ operation.**  
+  The operation now returns technical information about the accounting point and
+  the end user in addition to controllable units and technical resources.
+  The operation can also return HTTP 403 when the given arguments do not match
+  each other.
+
 ## 26.06.2025
 
 * **Specified the `data` field in the _Notice_ resource.**